+++ main.cc
@@ -1349,7 +1349,7 @@ To start the scylla server proper, simply invoke as: scylla server (or just scyl
             });
 
             raft_gr.start(raft::server_id{host_id.id}, std::ref(raft_address_map),
-                    std::ref(messaging), std::ref(gossiper), std::ref(fd)).get();
+                    std::ref(messaging), std::ref(fd)).get();
 
             // group0 client exists only on shard 0.
             // The client has to be created before `stop_raft` since during
@@ -1718,10 +1718,14 @@ To start the scylla server proper, simply invoke as: scylla server (or just scyl
                 group0_service.abort().get();
             });
 
-            load_address_map(sys_ks.local(), raft_address_map.local()).get();
-
             // Set up group0 service earlier since it is needed by group0 setup just below
             ss.local().set_group0(group0_service, raft_topology_change_enabled);
+
+            // Load address_map from system.peers and subscribe to gossiper events to keep it updated.
+            ss.local().init_address_map(raft_address_map.local()).get();
+            auto cancel_address_map_subscription = defer_verbose_shutdown("storage service uninit address map", [&ss] {
+                ss.local().uninit_address_map().get();
+            });
 
             // Need to make sure storage service does not use group0 before running group0_service.abort()
             auto stop_group0_usage_in_storage_service = defer_verbose_shutdown("group 0 usage in local storage", [&ss] {
